+++ src/server/master/index.ts
@@ -1,5 +1,5 @@
 import path from 'path';
-import { writeFileSync, copyFile, readdir, mkdir, readdirSync } from 'fs';
+import { writeFileSync, copyFile, readdir, mkdir, readdirSync, existsSync } from 'fs';
 import { promisify } from 'util';
 import master from './master';
 import creeveyApi, { CreeveyApi } from './api';
@@ -43,6 +43,7 @@ function readDirRecursive(dirPath: string): string[] {
 }
 
 function outputUnnecessaryImages(imagesDir: string, images: Set<string>): void {
+  if (!existsSync(imagesDir)) return;
   const unnecessaryImages = readDirRecursive(imagesDir)
     .map((imagePath) => path.posix.relative(imagesDir, imagePath))
     .filter((imagePath) => !images.has(imagePath));